+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.1.1]
+
+- Pin to newer dependencies of mlflow and algorithmia
+- Output new version of the model
+
 ## [0.1.0]
 
 - Initial release